requirements.txt
anyio==3.6.2
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
behave==1.2.6
black==23.3.0
cachier==1.5.3
certifi==2022.12.7
charset-normalizer==2.0.12
click==7.1
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
execnet==1.9.0
executing==1.2.0
Flask==2.2.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
logging==0.4.9.6
loguru==0.5.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
nest-asyncio==1.5.6
packaging==21.3
parse==1.19.0
parse-type==0.6.0
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pep8==1.7.1
pickleshare==0.7.5
platformdirs==3.2.0
pluggy==1.0.0
plumbum==1.7.1
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.0
pymongo==3.12.1
pyparsing==3.0.9
pypiwin32==223
pyTelegramBotAPI==4.11.0
pytest==7.3.0
pytest-cache==1.0
pytest-pep8==1.0.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rex==0.4
python-telegram-bot==20.2
pywin32==306
PyYAML==5.4.1
pyzmq==25.0.2
requests==2.26.0
rfc3986==1.5.0
search-letters==1.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
updater==0.3.2
urllib3==1.26.15
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
win32-setctime==1.1.0
